backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
python-jose>=3.3.0
passlib>=1.7.4
supabase>=1.0.3
openai>=1.1.1
python-multipart>=0.0.6
email-validator>=2.0.0
